refactor: fixed component named export, fixed nav bg style
