+++ src/components/atoms/star-badge.tsx
@@ -14,4 +14,4 @@ const StarBadge = ({rating}: Props) => {
   )
 }
 
-export default StarBadge
+export {StarBadge}
