+++ src/components/index.ts
@@ -1,12 +1,25 @@
-export * from './atoms/icons/facebook-icon'
-export * from './atoms/icons/github-icon'
-export * from './atoms/icons/google-icon'
-export * from './atoms/line'
-export * from './molecules/form/sign-in-form'
-export * from './molecules/form/sign-up-form'
-export * from './molecules/providers/auth-providers'
-export * from './ui/button'
-export * from './ui/checkbox'
-export * from './ui/form'
-export * from './ui/input'
-export * from './ui/label'
+export * from './atoms/icons/agent-icon';
+export * from './atoms/icons/facebook-icon';
+export * from './atoms/icons/github-icon';
+export * from './atoms/icons/google-icon';
+export * from './atoms/icons/nav-icon-button';
+export * from './atoms/line';
+export * from './atoms/star-badge';
+export * from './molecules/course-card/course-card';
+export * from './molecules/current-course-card/current-course-card';
+export * from './molecules/dashboard-card/dashboard-card';
+export * from './molecules/form/search-form';
+export * from './molecules/form/sign-in-form';
+export * from './molecules/form/sign-up-form';
+export * from './molecules/nav/nav-logo';
+export * from './molecules/nav/nav-menu';
+export * from './molecules/providers/auth-providers';
+export * from './organisms/NavBar/navbar';
+export * from './organisms/dashboard/dashboard';
+export * from './ui/button';
+export * from './ui/card';
+export * from './ui/checkbox';
+export * from './ui/form';
+export * from './ui/input';
+export * from './ui/label';
+export * from './ui/navigation-menu';
