+++ src/components/molecules/current-course-card/current-course-card.tsx
@@ -45,4 +45,4 @@ const CurrentCourseCard = ({ title, image }: CardProps) => {
   );
 };
 
-export default CurrentCourseCard;
+export {CurrentCourseCard};
