+++ src/components/molecules/nav/nav-menu.tsx
@@ -68,4 +68,4 @@ const NavMenu = () => {
   );
 };
 
-export default NavMenu;
+export {NavMenu};
